# readCon speedup